Move tests classes into their corresponding packages
